fix(mercury): tick semantics + channel split
Four review fixes on PR #131:

  1. **is_timed_out uses strict `>` against MAX_RETRIES.** A packet
     whose retransmit_count was bumped TO MAX_RETRIES on this tick now
     survives to be retransmitted; the channel is reaped only on the
     NEXT tick if that final retry still hasn't been ACKed within
     ACK_TIMEOUT_MS. Matches C++ Mercury (closes on the resend timer
     after the final retry, not on the retry itself).

  2. **Tick no longer eagerly touches `last_sent` for keepalives.** If
     the I/O layer drops a keepalive action, the next tick re-flags
     the same address rather than silently suppressing the keepalive
     for a full KEEPALIVE_INTERVAL_MS. Caller is now responsible for
     calling Channel::touch_sent after the bytes go on the wire.

  3. **Tick now sweeps stale fragment-reassembly state.** New
     `FRAGMENT_REASSEMBLY_TIMEOUT_MS` const (30s); tick calls
     Channel::cleanup_stale_fragments on every channel each pass.
     Without this, abandoned reassembly buffers would accumulate
     per channel until the channel itself died — and a chatty peer
     streaming fragments would keep the channel alive indefinitely.

  4. **Split channel.rs** (was 707 lines, past the 700-line hard cap).
     Promoted to channel/mod.rs (351 lines, impl) + channel/tests.rs
     (371 lines, all tests). No public API change.

Tests: +3 in nub::tests covering the new is_timed_out semantics
(survival on same-tick MAX_RETRIES hit, reap on next tick) and the
fragment-cleanup wire-up. The keepalive double-flag test was rewritten
to model the new contract: adjacent ticks WILL re-flag until the
caller calls touch_sent.

78 mercury tests pass.
